Convert <i> tag inside srt to proper tx3g italic style.

+++ mp4v2 wrapper/MP42Track.h
@@ -17,7 +17,7 @@
     NSString    *name;
     NSString    *language;
     BOOL        enabled;
-    int16_t     alternate_group;
+    uint64_t    alternate_group;
 
     BOOL    isEdited;
     BOOL    isDataEdited;
